pyproject.toml
[project]
name = "hooty"
dynamic = ["version"]
description = "Interactive AI coding assistant CLI"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
keywords = ["ai", "coding-assistant", "cli", "llm", "agno"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
]
dependencies = [
    "agno>=2.5.9",
    "typer>=0.24.1",
    "rich>=14.3.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0.47",
    "aiosqlite>=0.22.1",
    "prompt-toolkit>=3.0.52",
    "psycopg2-binary>=2.9",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.2",
    "beautifulsoup4",
    "httpx",
    "Pillow>=12.1.1",
    "filetype>=1.2",
]

[project.optional-dependencies]
anthropic = ["anthropic"]
azure = ["azure-ai-inference"]
azure-openai = ["openai"]
aws = ["boto3", "aioboto3"]
github = ["PyGithub"]
mcp = ["mcp"]
search = ["ddgs>=9.11.3"]
ollama = ["ollama"]
enterprise = ["cryptography>=43.0"]
all = ["hooty[anthropic,azure,azure-openai,aws,ollama,github,mcp,search,enterprise]"]
dev = [
    "pytest>=9.0",
    "pytest-cov",
    "ruff",
    "pyinstaller>=6.0",
    "pip-licenses",
]

[project.scripts]
hooty = "hooty.main:_cli_entry"

[project.urls]
Homepage = "https://github.com/niblings/hooty"
Repository = "https://github.com/niblings/hooty"
"Bug Tracker" = "https://github.com/niblings/hooty/issues"
Changelog = "https://github.com/niblings/hooty/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/hooty/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/hooty"]
artifacts = ["src/hooty/data/skills/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring API credentials",
]

[tool.ruff]
line-length = 120
src = ["src", "tests"]
